+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 - added: App distribution via Zealot
 - added: Add disableSurveyModal option
+- added: Handle KYC error responses from Bity
 - changed: Enable Bitpay payments for FIO names and domains
 - changed: Disable Home scene swap and FIO cards if configured
 - changed: Allow some chains to be disabled in env.json
 - changed: Add more metadata to zealot upload
 - fixed: Do not spin forever if the spam filter eliminates too many transactions.
 - fixed: FIO OBT data not showing in received transaction memos
+- fixed: Fix broken Bity message signing
 - fixed: Return underLimit error for Simplex for very small quotes
 - removed: Removed pull-to-search feature from wallet list and transaction list scenes